% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/clean.R
\name{acebypartnerquantity}
\alias{acebypartnerquantity}
\title{Replace quantity by quantity partner when the price variation in
recent periods was greater than that of the partner country for the same flow}
\usage{
acebypartnerquantity(dtf, choice)
}
\arguments{
\item{dtf}{data frame}

\item{choice}{a data frame of choice between reporter and partner}
}
\description{
Replace quantity by quantity partner when the price variation in
recent periods was greater than that of the partner country for the same flow
}

