% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{addconversionfactorandprice}
\alias{addconversionfactorandprice}
\title{Add unit price and conversion factor to each flow}
\usage{
addconversionfactorandprice(dtf)
}
\arguments{
\item{dtf}{data frame}
}
\description{
Calculate unit prices in current currency
Conversion factors in Quantity
}
