% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{adddot2logfile}
\alias{adddot2logfile}
\title{Add a dot at the end of a text file
while keeping a new line character at the end of the file
so that it is valid for \code{readLines}.}
\usage{
adddot2logfile(logfile)
}
\arguments{
\item{logfile}{character path to a log file}
}
\description{
Add a dot at the end of a text file
while keeping a new line character at the end of the file
so that it is valid for \code{readLines}.
}
