% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{addmissingmirrorflow}
\alias{addmissingmirrorflow}
\title{Add the partner flow when quantity is missing}
\usage{
addmissingmirrorflow(dtf)
}
\arguments{
\item{dtf}{a data frame containing world trade flows for one product}
}
\description{
When a flow does not have a mirror flow.
}

