% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{addpartnerflow}
\alias{addpartnerflow}
\title{Add volume and value of the partner flow}
\usage{
addpartnerflow(dtf)
}
\arguments{
\item{dtf}{data frame containing trade flow data.
With column names following the efi convention.}
}
\description{
For the same country, period, item, add the corresponding partner flow.
}

