% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{addregion}
\alias{addregion}
\title{Add regionreporter and regionpartner}
\usage{
addregion(dtf, regioncolumn = "region")
}
\arguments{
\item{dtf}{data frame of trade flows}

\item{regioncolumn}{column containing region names}
}
\description{
Add regionreporter and regionpartner
}

