% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{calculatediscrepancies}
\alias{calculatediscrepancies}
\title{Calculate discrepancies}
\usage{
calculatediscrepancies(dtf)
}
\arguments{
\item{dtf}{data frame reshaped in long format.
containing trade flow data.
With column names following the efi convention.}
}
\description{
The plan is to use this as needed before plotting
but not to store it in the enddata
as it is for plotting with ggplot,
it might work best on a reshaped data frame
}

