% Generated by roxygen2 (4.0.2): do not edit by hand
\name{changecolumntype}
\alias{changecolumntype}
\title{Change some of the column types}
\usage{
changecolumntype(dtf)
}
\arguments{
\item{dtf}{data frame}
}
\description{
Change to factors for easy plotting
 Change to large int or to floating point for some computations
}

