% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{changeflowmessage}
\alias{changeflowmessage}
\title{A function to calculate the percentage change on world imports and exports}
\usage{
changeflowmessage(dtf0, dtf1, import = "Import", export = "Export")
}
\arguments{
\item{dtf0}{a data frame of trade flows before the modification}

\item{dtf1}{a data frame of modified trade flows}

\item{import}{code indicating import data, can be "Import" for Comtrade and
1 for Comext}

\item{export}{code indicating export data, can be "Export" for Comtrade and
2 for Comext

Note: Add a function changeflowmessage(dtf, dtfsplit) which will return a character
string about both trade flows. This function could be defined in the main scope
and reused in other parts of the workflow. Or maybe not because the quantity is
not empty for addpartner flow and shaveprice.}
}
\description{
A function to calculate the percentage change on world imports and exports
}

