% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/clean.R
\name{choosereporterorpartner}
\alias{choosereporterorpartner}
\title{Compare reporter and partner flow}
\usage{
choosereporterorpartner(dtf, periodbegin = 2009, periodend = 2013,
  sdratiolimit = 0.8)
}
\arguments{
\item{dtf}{data frame}

\item{periodbegin}{change this to global parameter}

\item{periodend}{change this to global parameter}
}
\description{
The data frame outcome of this function
will be use by replacebypartnerquantity()
to decide which of the reporter or partner flow to favor
}

