% Generated by roxygen2 (4.0.2): do not edit by hand
\name{clean}
\alias{clean}
\alias{cleantodb}
\title{Combine cleaning functions}
\usage{
clean(dtf, writetodb = FALSE, shaveprice = FALSE,
  deleteextracolumns = TRUE)

cleantodb(..., writetodb = TRUE)
}
\arguments{
\item{dtf}{data frame}

\item{deleteextracolumns}{when TRUE, keep only columns from
column_names$validated_flow}

\item{...}{arguments passed to \code{\link{clean}()}
from \code{\link{cleantodb}()}}

\item{write2db}{logical}

\item{write2db}{logical TRUE to write the result in the configured database}
}
\description{
Use unit prices and conversion factors to complete missing quantity data.
Handle unit prices that are out of bound.
Add quantity estimates reported by the trade partner.
If writetodb ==TRUE, the outcome will be written to the default database.
Otherwise the outcome is returned as a data frame.
}

