% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleancomext.R
\name{cleancomextmonthly1product}
\alias{cleancomextmonthly1product}
\title{Clean Comext Monthly data}
\usage{
cleancomextmonthly1product(RMySQLcon, productanalysed, tablearchive,
  tablerecent, tableprice, tablepricew, tablecv)
}
\arguments{
\item{RMySQLcon}{database connection object created by RMySQL \code{\link[DBI]{dbConnect}}}

\item{tablearchive}{character name of a monthly archive table}

\item{tablerecent}{character name of a monthly recent table}

\item{tableprice}{character name of a price table}

\item{tablecv}{character name of a conversion factor table}
}
\description{
Clean Comext Monthly data
}
\examples{
\dontrun{ # Clean product and country codes
# Connect to the database
con <- RMySQL::dbConnect(RMySQL::MySQL(), dbname = "test")
# Clean product 44079910

# Disconnect from the database
RMySQL::dbDisconnect(con)
}
}

