% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/expertexcel.R
\name{cleandb2excel}
\alias{cleandb2excel}
\title{Clean from a database table to Excel, for expert analysis.}
\usage{
cleandb2excel(productcode, tableread)
}
\arguments{
\item{productcode}{code of a product}

\item{tableread}{names of the database table to read}
}
\description{
Clean from a database table to Excel, for expert analysis.
}

