% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clean.R
\name{cleandbproduct}
\alias{cleandbproduct}
\title{Write flows into the database table validated_flow}
\usage{
cleandbproduct(productcode, tableread, tablewrite, ...)
}
\arguments{
\item{tableread}{name of the table to read from}

\item{tablewrite}{name of the table to write to (all rows for productcode
will be deleted in this table before writing)}

\item{...}{further arguments passed to the clean function}

\item{productcode_}{code of the product trade flows to be validated}
}
\value{
TRUE if write to db succesded, otherwise return FALSE
}
\description{
For one product (at the 6 digit level),
read and clean all trade flows, then
Write flows into the database table(s) validated_flow
updates will be done on a product basis,
The function will:
\enumerate{
 \item{Read all flows having the given productcode in tableread}
 \item{Delete all flows having the given productcode
 in tablewrite (between all reporter and partner countries in all years)}
 \item{Use  \code{\link{clean}()} to clean the data frame}
 \item{Write validated flows to tablewrite}
}
}

