% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{cleanmonthly}
\alias{cleanmonthly}
\title{Clean monthly trade flows}
\usage{
cleanmonthly(dtfmonthly, dtfyearly, geoaggregation = "region",
  replacebypartnerquantity = TRUE, shaveprice = TRUE,
  outputalltables = FALSE)
}
\arguments{
\item{dtfmonthly}{data frame of monthly trade flows}
}
\description{
The function uses regional conversion factors
and unit prices generated from the yearly data frame.
}

