% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_column_names.R
\docType{data}
\name{column_names}
\alias{column_names}
\title{Column names of data sources}
\format{a data frame with 4 columns corresponding to each data source}
\usage{
column_names
}
\description{
A match table between the column names of
different trade data sources.
}
\section{Data sources}{


\itemize{
\item \code{comtrade}: Column names from
the  un comtrade api \url{http://comtrade.un.org/data/doc/api/}.
\item \code{faostat}
\item \code{eurostat}
\item \code{tradeflows} : column names of the current project
}
}
\examples{
column_names
}

