% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/report.R
\name{createalloverviewreports}
\alias{createalloverviewreports}
\title{Create overview reports for all countries available}
\usage{
createalloverviewreports(tableread = "validated_flow_yearly",
  outputpath = "reports/overview")
}
\arguments{
\item{tableread}{character name of a database table}

\item{path}{where reports will be located}
}
\description{
Create overview reports for all countries available
}
\examples{
\dontrun{
createalloverviewreports("validated_flow_yearly", "/tmp")
}
}

