% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{createcountryindex}
\alias{createcountryindex}
\title{Generate a list of reports}
\usage{
createcountryindex(countries, template = "countryindex.Rmd",
  inputpath = system.file("templates", package = "tradeflows"),
  outputdir = "reports/overview", filename = "countryindex.html",
  encoding = "UTF-8")
}
\arguments{
\item{template}{name of the template file, including .Rmd extension}

\item{inputpath}{path of the template, defaults to package internal path}

\item{outputdir}{name of the output directory relative to getwd()}

\item{filename}{name of the output file}

\item{encoding, }{encoding of the template file}
}
\description{
Generate a list of reports
}
\examples{
\dontrun{
countries <- data_frame(reporter = c("Finland", "France"))
createcountryindex(countries)
}
}

