% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/report.R
\name{createcountryreport}
\alias{createcountryreport}
\title{Create reports for countries
Illustrating major trade flows.}
\usage{
createcountryreport(countryinreport)
}
\description{
Create reports for countries
Illustrating major trade flows.
}

