% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{createdbstructure}
\alias{createdbstructure}
\title{Creates a database structure based on a sql dump file}
\usage{
createdbstructure(sqlfile, dbname = "tradeflows",
  sqlfolder = system.file("config", package = "tradeflows", mustWork = TRUE),
  messageonly = FALSE, verbose = TRUE)
}
\arguments{
\item{sqlfile}{character name of an SQL file}

\item{dbname}{character name of a database}

\item{sqlfolder}{character path to the folder where the sqlfile is located
(defaults to the package config directory)}

\item{messageonly}{boolean if TRUE prints only a message and doesn't
transfer the database structure}
}
\description{
Load a .sql file containing a database structure.
}
\details{
The default .sql file is storred in the package configuration folder.
}
\examples{
# Load new database structures in the test database
# Load a database structure designed to contain raw data
createdbstructure(sqlfile = "raw_comext.sql", dbname = "test")
# Load a database structure designed to contain validated data
createdbstructure(sqlfile = "val_comext.sql", dbname = "test")
}

