% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/report.R
\name{creatediscrepancyreport}
\alias{creatediscrepancyreport}
\title{Create a discrepancy report}
\usage{
creatediscrepancyreport(productcode_, reporter_, beginyear = 0,
  endyear = 9999, template = "discrepancy.Rmd",
  outputdir = "reports/discrepancies", tableread = "raw_flow_yearly",
  toc = FALSE, ...)
}
\arguments{
\item{reporter_}{one single country name}

\item{...}{arguments passed to \code{\link{createreport}()}}

\item{productcode}{vector of product codes}
}
\description{
Create a discrepancy report
}
\examples{
\dontrun{
creatediscrepancyreport(440799, "Cameroon", outputdir = "reports/discrepancies")
creatediscrepancyreport(440710, "Thailand", beginyear = 2010, endyear = 2012)
}
}

