% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/report.R
\name{createproductreport}
\alias{createproductreport}
\title{Create data completeness report for comtrade data}
\usage{
createproductreport(tfdata, inputpath = system.file("templates", package =
  "tradeflows"), template, outputdir = "reports", keep_tex = FALSE)
}
\arguments{
\item{tfdata}{a dataframe containing tradeflows data}

\item{inputpath}{path of the template, defaults to package internal path}

\item{template}{name of the template file, including .Rmd extension}

\item{outputdir}{name of the output directory relative to getwd()}
}
\description{
This function will generate reports for any template that is
product specific.  Giving information about all world trade
flows for one product.
}

