% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/template.R
\name{createreport}
\alias{createcompletenessreport}
\alias{createreport}
\title{Create data completeness report for comtrade data}
\usage{
createreport(rawdata, products = "all", inputpath = system.file("templates",
  package = "tradeflows"), template, outputdir = "reports")

createcompletenessreport(..., template = "completeness.Rmd")
}
\arguments{
\item{rawdata}{a dataframe of tradeflows data}

\item{products}{a vector of product codes}

\item{inputpath}{path of the template, defaults to package internal path}

\item{template}{name of the template file, including .Rmd extension}

\item{outputdir}{name of the output directory relative to getwd()}

\item{...}{arguments passed to \code{\link{createreport}()}
from \code{createcompletenessreport}()}

\item{template}{name of the template file}
}
\description{
You may want to generalise this function, for all template that are
product specific.
}

