% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{createreportfromdb}
\alias{createreportfromdb}
\title{Create report from the database}
\usage{
createreportfromdb(tableread, products = "all", ...)
}
\arguments{
\item{products}{a vector of product codes}

\item{...}{arguments passed to \code{\link{createreport}()}
from \code{createcompletenessreport}()}

\item{template}{name of the template file}
}
\description{
Create a report from the database
}
\details{
For a given vector of product codes, create reports using a given template.
}
