% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/database.R
\name{deletedbproduct}
\alias{deletedbproduct}
\title{Delete validated data for one product}
\usage{
deletedbproduct(productcode, tabledelete)
}
\arguments{
\item{product}{name}
}
\description{
Delete validated data for one product
}

