% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/report.R
\name{description}
\alias{description}
\title{Product description}
\usage{
description(productcodes)
}
\arguments{
\item{productcodes}{numeric vector of product codes}
}
\description{
To be used in a markdown formated document.
Use:
description(c(440799, 440795))
}

