% Generated by roxygen2 (4.0.2): do not edit by hand
\name{estimatequantity}
\alias{estimatequantity}
\title{Estimate quantity}
\usage{
estimatequantity(dtf, price, conversionfactor, shaveprice = TRUE)
}
\arguments{
\item{dtf}{data frame}

\item{price}{a data frame of unit prices which can be merge with dtf}

\item{conversionfactor}{a data frame of conversion factors which can be
merged with dtf}

\item{shave}{logical, TRUE when upper and lower bound on prices have to be
checked and quantity replaced accordingly}
}
\description{
Estimate missing quantity
For each trade flow in the given data frame,
compute quantity_cf from the weight using a conversion factor and
comput quantity_up from the trade value using a unit price.
These columns will be available in the data frame output of this function,
but they will not be saved in the validated database.
Then split the data frame
complete missing quantity from weight when available or from the
tradevalue when the weight is not available
Check unit price
Check unit price of the trade flow, against a table of reference unit prices
Check upper and lower bounds on price.
In general NA values should be avoided for the
 upper and lower bounds on prices
}

