% Generated by roxygen2 (4.0.2): do not edit by hand
\name{extractconversionfactors}
\alias{extractconversionfactors}
\title{Extract median converion factor at a given geographical aggregation level}
\usage{
extractconversionfactors(dtf, geoaggregation = "World",
  excludeqwestimates = FALSE)
}
\arguments{
\item{dtf}{a dataframe containing conversion factors}

\item{geoaggregation}{a character string specifying the regional aggregation
level, a column name in the reportercomtrade table}

\item{excludeqwestimates}{logical TRUE when comtrade quantity and weight
estimates have to be excluded}
}
\description{
Extract median converion factors for the whole world (default)
}

