% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/expertexcel.R
\name{extractflags}
\alias{extractflags}
\title{Extract information from flag estimates}
\usage{
extractflags(dtf)
}
\arguments{
\item{dtf}{a cleaned data frame containing trade flows}
}
\description{
Extract information from flag estimates
}

