% Generated by roxygen2 (4.0.2): do not edit by hand
\name{extractprices}
\alias{extractprices}
\title{Extract median prices at a given geographical aggregation level}
\usage{
extractprices(dtf, geoaggregation = "regionreporter",
  excludeqestimates = FALSE)
}
\arguments{
\item{dtf}{a dataframe containing already available prices}

\item{geoaggregation}{a character vector specifying the regional aggregation
level, a column name in the reportercomtrade table}

\item{excludeqestimates}{logical TRUE when comtrade quantity estimates
have to be excluded}
}
\description{
Extract median prices by region (default) or at another
geographical aggregation level (subregion)
}

