% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clean.R
\name{filterworldeu28}
\alias{filterworldeu28}
\title{Filter world and EU28}
\usage{
filterworldeu28(dtf)
}
\arguments{
\item{dtf}{data frame containing comtrade trade flows}
}
\description{
Remove EU28 reporter
Remove EU-28 and World partner
}

