% Generated by roxygen2 (4.0.2): do not edit by hand
\name{findduplicatedflows}
\alias{findduplicatedflows}
\title{Return a dataframe of duplicated flows}
\usage{
findduplicatedflows(dtf)
}
\arguments{
\item{dtf}{a dataframe contiaining trade flows, the funciton
tests for duplicated reportercode, partnercode, productcode, flow, year}
}
\description{
Return a dataframe of duplicated flows
}

