% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{joinpricecvfbounds}
\alias{joinpricecvfbounds}
\title{Join price and conversion factor tables to the data frame}
\usage{
joinpricecvfbounds(dtf, price, conversionfactor)
}
\arguments{
\item{dtf}{data frame}

\item{price}{a data frame of unit prices which can be merge with dtf}

\item{conversionfactor}{a data frame of conversion factors which can be
merged with dtf}
}
\description{
Merge a price table and a conversion factor table to the main dataframe.
The price table was generated by the extractprices function and
the conversion factor table was generated by the extractconversionfactors function.
}

