% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{lengthunique}
\alias{lengthunique}
\title{Return the length of unique combination of given column names}
\usage{
lengthunique(dtf, ...)
}
\arguments{
\item{dtf}{data frame}

\item{...}{further arguments passed to \code{\link{select_}()}}
}
\description{
Return the length of unique combination of given column names
}
\examples{
lengthunique(airquality)
lengthunique(airquality, "Month", "Day")
lengthunique(airquality, "Month")
lengthunique(airquality, "Day")
}

