% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleancomext.R
\name{loadcomext1product}
\alias{loadcomext1product}
\title{Load all archive and recent Comext data for one product}
\usage{
loadcomext1product(RMariaDBcon, productanalysed, tablearchive, tablerecent,
  tableunit = "vld_comext_unit")
}
\value{
a  data frame containing Comext trade flows for the given product
}
\description{
The function will use the database connector provided as
RMariaDBcon argument.
It will load the product from an archive table and a recent
table. The function will combine the archive table and
the recent table in one dataframe and will return a dataframe.
}
\examples{
\dontrun{ # Clean product and country codes
# Connect to the database
con <- RMariaDB::dbConnect(RMariaDB::MariaDB(), dbname = "tradeflows")
# Load data for product 44079910
dtf <- loadcomext1product(con,
                          productanalysed = "44071091",
                          tablearchive = "raw_comext_monthly_2016S1",
                          tablerecent = "raw_comext_monthly_201709")
head(dtf)
unique(dtf$year)
# Disconnect from the database
RMariaDB::dbDisconnect(con)
}
}
