% Generated by roxygen2 (4.0.2): do not edit by hand
\name{loadcomtrade_all_countries}
\alias{loadcomtrade_all_countries}
\title{load comtrade data for all countries, for a vector of product codes
this could be moved later.}
\usage{
loadcomtrade_all_countries(productcode)
}
\arguments{
\item{productcode}{a vector of product codes}
}
\description{
load comtrade data for all countries, for a vector of product codes
this could be moved later.
}

