% Generated by roxygen2 (4.0.2): do not edit by hand
\name{loadcomtrade_bycode}
\alias{loadcomtrade_bycode}
\title{load raw data from the COMTRADE API and other sources}
\usage{
loadcomtrade_bycode(productcode, reportercode, year, px = "HS", max = 50000)
}
\arguments{
\item{productcode}{code of the product}

\item{reportercode}{geographical area or country code}

\item{year}{a string of year separated by commas}

\item{px}{Trade data classification scheme}

\item{max}{maximum records returned}
}
\value{
a dataframe
}
\description{
load a JSON file from the comtrade API in the temp directory
The API is documented at http://comtrade.un.org/data/doc/api/
Converts the file to a dataframe
}

