% Generated by roxygen2 (4.0.2): do not edit by hand
\name{loadcomtradeallreporters}
\alias{loadcomtradeallreporters}
\title{Load flows for all countries, in all directions available in comtrade
for a given product.}
\usage{
loadcomtradeallreporters(productcode, year = "recent")
}
\arguments{
\item{productcode}{a vector of product codes}

\item{year}{a vector of years (maximum 5), or the chain of character "recent"}
}
\description{
Load flows for all countries, in all directions available in comtrade
for a given product.
}

