% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{loadcomtradebycode}
\alias{loadcomtradebycode}
\title{load raw data from the COMTRADE API}
\usage{
loadcomtradebycode(productcode, reportercode, year, freq = "A",
  px = "HS", max = 50000, logfile = FALSE)
}
\arguments{
\item{productcode}{code of the product}

\item{reportercode}{geographical area or country code}

\item{year}{a vector of years (maximum 5), or the chain of character "recent"}

\item{freq}{the frequency: "A" Annual, "M" Monthly}

\item{px}{Trade data classification scheme}

\item{max}{maximum records returned}
}
\value{
a dataframe
}
\description{
load a JSON file from the comtrade API in the temp directory
The API is documented at \url{http://comtrade.un.org/data/doc/api/}
Converts the file to a dataframe
Further information for developers is available at
\url{https://comtrade.un.org/Data/Doc/API}
}
\examples{
\dontrun{
library(tradeflows)
# Load trade data from Comtrade
# "other sawnwood" trade France in 2018
loadcomtradebycode(440799, 251, 2018)
}
}
