% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/load.R
\name{loadcomtradewithpause}
\alias{loadcomtradewithpause}
\title{Load flows for all countries, with a 1 hour pause between products}
\usage{
loadcomtradewithpause(products, year = "recent", pause = 3601)
}
\arguments{
\item{year}{a vector of years (maximum 5), or the chain of character "recent"}

\item{pause}{length of the pause in seconds}

\item{productcode}{a vector of product codes}
}
\description{
Load flows in all directions available in comtrade
for a given vector of product codes.
}

