% Generated by roxygen2 (4.0.2): do not edit by hand
\name{loadrawdata}
\alias{loadrawdata}
\title{Load raw data form the database}
\usage{
loadrawdata(productcode_, table = "raw_trade_data")
}
\arguments{
\item{productcode}{the code of on product}
}
\description{
Connect to the raw database and load all tradeflows
for the given product code,
for all years, in all directions, for all reporter and all partners.
}

