% Generated by roxygen2 (4.0.2): do not edit by hand
\name{nrowinDB}
\alias{nrowinDB}
\title{Number of rows in a database tables}
\usage{
nrowinDB(table)

nrowinDB(table)
}
\arguments{
\item{table}{name of a table}

\item{table}{name of a table}
}
\description{
Number of rows in a database tables

Number of rows in a database tables
}

