% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/database.R
\name{readdbproduct}
\alias{readdbproduct}
\title{Load trade flows for one product from a database table of choice}
\usage{
readdbproduct(productcode_, tableread)
}
\arguments{
\item{productcode_}{the code of a product}

\item{tableread}{the database table to read from}
}
\description{
Connect to the raw database and load all tradeflows
for the given product code,
for all years, in all directions, for all reporter and all partners.
}

