% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/database.R
\name{readdbtbl}
\alias{readdbtbl}
\title{Create a dplyr connector to MySQL}
\usage{
readdbtbl(tableread)
}
\arguments{
\item{tableread}{name of the database table to read}
}
\value{
a dplyr tbl object
}
\description{
Return a dplyr tbl object for the given MySQL database table
tbl objects allow lazy operations. See src_mysql.
}
\examples{
\dontrun{
readdbtbl("raw_flow_yearly") \%>\%
    group_by(productcode) \%>\%
    summarise(nrow = n()) \%>\%
    arrange(nrow) \%>\%
    kable
}
}

