% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clean.R
\name{removeduplicatedflows}
\alias{removeduplicatedflows}
\title{Remove duplicated flows}
\usage{
removeduplicatedflows(dtf)
}
\arguments{
\item{dtf}{a dataframe contiaining trade flows,}
}
\description{
Print the number of duplicated entries and remove them if needed.
The function tests for all duplicated columns
(contrary to findduplicatedflows() which looks only for some columns).
This might be an issue if 2 flows have duplicated
reportercode, partnercode, productcode, flow, period but
their quantity column is different.
}

