% Generated by roxygen2 (4.0.2): do not edit by hand
\name{savedtf2excel}
\alias{savedtf2excel}
\title{Save a data frame to the Excel format, with column names}
\usage{
savedtf2excel(dtf, filename, path = "data-raw/")
}
\arguments{
\item{dtf}{data frame}

\item{filename}{name of the excel file}

\item{path}{folder where the file will be saved

Todo: could use deparse(substitute(dtf)) to create a filename from the objects name
this might not be needed,
as the ?xlsx explains
write.xlsx(USArrests, file="file.xlsx")}
}
\description{
This function is not needed use xlsx::write.xlsx() directly
But developing the function below might be usefull
}

