% Generated by roxygen2 (4.0.2): do not edit by hand
\name{savelistdtf2excel}
\alias{savelistdtf2excel}
\title{Save a list of data frames to Excel format}
\usage{
savelistdtf2excel(ldtf, filename, path)
}
\arguments{
\item{ldtf}{Named list of Data Frames}
}
\description{
creates an excel file containing sheets for each data frame in the list.
sheet names are created from the data frame names in the list.
Data frames have to be named.
}
\details{
modifiy this function or create a new function that
      saves a list of data frames to excel sheets could test if dtf is a list
}

