% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/database.R
\name{selectbyid}
\alias{selectbyid}
\title{Select rows by id}
\usage{
selectbyid(tableread, id)
}
\arguments{
\item{id}{a vector of id}

\item{table}{name of a table}
}
\description{
Select rows by id
}

