% Generated by roxygen2 (4.0.2): do not edit by hand
\name{setdatabaseconfig}
\alias{setdatabaseconfig}
\title{Sets database connection parameters as a global R option tradeflowsDB}
\usage{
setdatabaseconfig(reload = FALSE)
}
\arguments{
\item{reload}{logical}
}
\description{
Database connection parameters
Are storred in the installed version of the package
Run this command to find the location of the configuration file on your machine:
system.file("config/databaseconf.R", package="tradeflows").
Use reload=TRUE to force reloading the file after modification.
}

