% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/clean.R
\name{shaveprice}
\alias{shaveprice}
\title{Change quantities when unit prices are too high or too low}
\usage{
shaveprice(dtf)
}
\arguments{
\item{dtf}{data frame}
}
\description{
Upper and lower bound on prices have to be
checked and quantity replaced accordingly
This function comes after estimatequantity()
and after addpartnerflow()
check this order in the source of the clean() function.
Check unit price for all flows
recalculate unit price based on quantity estimate
Those which have been estimated from tradevalue will have correct prices
those which have been estimated from weight might be changed as well
}

