% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{swapreporterpartner}
\alias{swapreporterpartner}
\title{Rename reporter to partner and partner to reporter}
\usage{
swapreporterpartner(dtf, column = c("reportercode", "partnercode",
  "productcode", "flow", "period", "quantity", "tradevalue"))
}
\arguments{
\item{dtf}{data frame containing trade flow data}

\item{column}{column names to select, NULL if all columns have to be selected
Selected column names should at least include the identity columns which will be used
used for the merge}
}
\description{
Rename reporter to partner and partner to reporter
}
