% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{tradeflows-deprecated}
\alias{tradeflows-deprecated}
\title{Deprecated functions in the tradeflows package}
\description{
Deprecated functions in the tradeflows package
}
\section{Details}{

\tabular{rl}{
  \code{\link{readdbtbl}} \tab should now be replaced by \code{dplyr::\link{tbl}}\cr
}
}

