% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/report.R
\name{trytocreateoverviewreports}
\alias{trytocreateoverviewreports}
\title{Fault tolerant report creation for the server}
\usage{
trytocreateoverviewreports(reporter, tableread = "validated_flow_yearly",
  outputpath = "reports/overview")
}
\arguments{
\item{reporter}{character containing one reporter name}
}
\description{
Can be used with lapply to generate reports for a list of countries.
}
\examples{
\dontrun{
countries <- c("Finland","France")
lapply(countries, trytocreateoverviewreports)
}
}

