% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{writedbproduct}
\alias{writedbproduct}
\title{Write validated flows for one product into the database}
\usage{
writedbproduct(dtf, tablewrite)
}
\arguments{
\item{dtf}{a data frame containing the data to be inserted}

\item{tablewrite}{character vector of a database table to write to}
}
\description{
Write data into the DB
}
