% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{writeerror2log}
\alias{writeerror2log}
\title{Write errors to a log file}
\usage{
writeerror2log(condition, logfile, optionaltext = NULL)
}
\arguments{
\item{condition}{warning or error condition as returned by \link{tryCatch}}

\item{logfile}{character path to a log file}

\item{optionaltext}{character additional text added after the timestamp}
}
\description{
Write errors to a log file
}
\examples{
# Create a log file name
logfileexample <- file.path(tempdir(),"log.txt")
# Catch an error
 tryCatch({
     stop("There is an error in this exampe.")
 }, error = function(errorcondition){
            writeerror2log(errorcondition, logfileexample, "code: 123456")
 })
 # Read the log file
readLines(logfileexample)
}
