% Generated by roxygen2 (4.0.2): do not edit by hand
\name{addconversionfactorandprice}
\alias{addconversionfactorandprice}
\title{Add unit price and conversion factor to each flow}
\usage{
addconversionfactorandprice(dtf)
}
\arguments{
\item{dtf}{data frame}
}
\description{
Calculate unit prices in current currency
Conversion factors in Quantity
}

