% Generated by roxygen2 (4.0.2): do not edit by hand
\name{addpartnerflow}
\alias{addpartnerflow}
\title{Add volume and value of the partner flow}
\usage{
addpartnerflow(dtf)
}
\arguments{
\item{dtf}{data frame containing trade flow data.
With column names following the efi convention.}
}
\description{
For the same country, year, item, add the corresponding partner flow.
}

