% Generated by roxygen2 (4.0.2): do not edit by hand
\name{clean}
\alias{clean}
\title{Calculate volume based on tradevalue
Combine all cleaning functions}
\usage{
clean(dtf, write2db = FALSE)
}
\arguments{
\item{dtf}{data frame}

\item{write2db}{logical TRUE to write the result in the database}
}
\description{
Calculate volume based on tradevalue
Combine all cleaning functions
}

