% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/clean.R
\name{clean}
\alias{clean}
\title{Combine cleaning functions}
\usage{
clean(dtf, replacebypartnerquantity = TRUE, shaveprice = TRUE,
  deleteextracolumns = FALSE, outputalltables = FALSE)
}
\arguments{
\item{dtf}{data frame}

\item{replacebypartnerquantity,}{when TRUE call replacebypartnerquantity()}

\item{shaveprice}{when TRUE, call shaveprice()}

\item{deleteextracolumns}{when TRUE, keep only columns from
tradeflows::column_names$validated_flow}

\item{outputalltables}{when TRUE return a list of data frames
containing the cleaned dataset and all tables used to compute it
conversion factors, regional unit prices, choice between
reporter and partner flows}
}
\description{
Use unit prices and conversion factors to complete missing quantity data.
Handle unit prices that are out of bound.
Add quantity estimates reported by the trade partner.
The outcome is returned as a data frame.
}

