% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/expertexcel.R
\name{clean2excel}
\alias{clean2excel}
\title{Clean and export to Excel for the expert users}
\usage{
clean2excel(dtf, filenamestart)
}
\arguments{
\item{dtf}{data frame}

\item{filename}{path and name of the output files,
without extensions}
}
\description{
Keep price, conversion factor andflow choice in a separate table
for further analysis
}

