% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/expertexcel.R
\name{clean2excel}
\alias{clean2excel}
\alias{cleandb2excel}
\alias{cleanrdata2excel}
\title{Clean and export price and conversion factor to Excel}
\usage{
clean2excel(dtf, file, path = "data-raw/excel", returnresults = FALSE, ...)

cleandb2excel(productcode, file = paste0(productcode, ".xlsx"),
  tableread = "raw_flow_yearly", ...)

cleanrdata2excel(rawfilename, ...)
}
\arguments{
\item{dtf}{data frame}

\item{file}{a character string naming an Excel file, should end by ".xlsx"}

\item{path}{a character string specifying where the excel file will be saved}

\item{returnresults}{TRUE if the list of results should be returned
usefull for the function cleanrdata2excel.}

\item{...}{further arguments passed to \code{\link{clean()}}
without extensions}

\item{productcode}{code of a product}

\item{tableread}{names of the database table to read}

\item{rawfile}{name of a .RData file containing raw trade flows}
}
\description{
Save price, conversion factor and flow choice in a separate table
for further analysis.
Use the argument file to specify where the Excel file will be located.

\code{cleandb2excel} cleans from a database table to Excel.

\code{cleanrdata2excel} cleans from a raw data file to Excel and
write the cleaned dataset to a csv file.
}

