% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleancomext.R
\name{cleancode}
\alias{cleancode}
\title{Extract a list of unique, most recent product and country codes from Comext}
\usage{
cleancode(RMySQLcon, tableread, tablewrite, codevariable)
}
\arguments{
\item{RMySQLcon}{database connection object created by \code{RMySQL::\link{dbConnect}}}

\item{tableread}{character name of the table to read from}

\item{tablewrite}{character name of the table to write to}

\item{codevariable}{unquoted code variable (à la dplyr verbs)}
}
\value{
TRUE on success
The output is actually a database table containing the cleaned codes.
}
\description{
Takes a raw codes table from comext, select codes which have the most
recent \code{datestart} and make sure they are unique.
}
\examples{
\dontrun{
con <- RMySQL::dbConnect(RMySQL::MySQL(), dbname = "test")
# Write dummy codes to the database
raw_code <- data.frame(code = c(4L, 4L), datestart = c(1L, 2L))
RMySQL::dbWriteTable(con, "raw_code", raw_code, row.names = FALSE, overwrite = TRUE)
# Clean them
cleancode(con, tableread = "raw_code", tablewrite = "val_code", codevariable = "code")
# Clean real comext codes


on.exit(RMySQL::dbDisconnect(con))
}
}

