% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleancomext.R
\name{cleancomextmonthly1product}
\alias{cleancomextmonthly}
\alias{cleancomextmonthly1product}
\title{Clean Comext Monthly data}
\usage{
cleancomextmonthly1product(RMySQLcon, productanalysed, tablearchive,
  tablerecent, tablewrite, tableprice = "vld_comext_price",
  tablepricew = "vld_comext_pricew", tablecv = "vld_comext_cv")

cleancomextmonthly(RMySQLcon, productanalysed, tablearchive, tablerecent,
  tablewrite, ...)
}
\arguments{
\item{RMySQLcon}{database connection object created by RMySQL \code{\link[DBI]{dbConnect}}}

\item{tablearchive}{character name of a monthly archive table}

\item{tablerecent}{character name of a monthly recent table}

\item{tableprice}{character name of a price table}

\item{tablecv}{character name of a conversion factor table}

\item{...}{parameters passed to cleancomextmonthly1product}
}
\value{
\code{cleancomextmonthly1product} invisibly returns a data frame
with all columns generated during the cleaning process.
}
\description{
Clean Comext Monthly data
}
\examples{
\dontrun{ # Clean product and country codes
# Connect to the database
con <- RMySQL::dbConnect(RMySQL::MySQL(), dbname = "test")
# Clean product 44079910
cleancomextmonthly1product(con ,
                           productanalysed = "44071091",
                           tablearchive = "raw_comext_monthly_2016S1",
                           tablerecent = "raw_comext_monthly_201709",
                           tablewrite = "vld_comext_monthly",
                           tableprice = "vld_comext_price",
                           tablepricew = "vld_comext_pricew",
                           tablecv = "vld_comext_cv")
cleancomextmonthly(con ,
                   tablearchive = "raw_comext_monthly_2016S1",
                   tablerecent = "raw_comext_monthly_201709",
                   tablewrite = "vld_comext_monthly",
                   tableprice = "vld_comext_price",
                   tablepricew = "vld_comext_pricew",
                   tablecv = "vld_comext_cv")
# Disconnect from the database
RMySQL::dbDisconnect(con)
}
}

