% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/expertexcel.R
\name{cleandb2excel}
\alias{cleandb2excel}
\title{Clean from a database table to Excel, for expert analysis.}
\usage{
cleandb2excel(productcode, tableread = "raw_flow_yearly", ...)
}
\arguments{
\item{productcode}{code of a product}

\item{tableread}{names of the database table to read}

\item{...}{further arguments passed to clean()}
}
\description{
Use the argument filenamestart to specify where the Excel file will be located.
}

