% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{createcompletenessreport}
\alias{createcompletenessreport}
\title{Create the completeness report}
\usage{
createcompletenessreport(productcode_, beginyear = 0, endyear = 9999,
  template = "completeness.Rmd", outputdir = "reports/completeness",
  tableread = "raw_flow_yearly", toc = TRUE, ...)
}
\arguments{
\item{productcode_}{vector of product codes}

\item{...}{arguments passed to \code{\link{createreport}()}}
}
\description{
Create the completeness report
}
\examples{
\dontrun{
createcompletenessreport(440799, beginyear = 2010, endyear = 2011)
}
}
