% Generated by roxygen2 (4.0.2): do not edit by hand
\name{createcompletenessreport}
\alias{createcompletenessreport}
\title{Create data completeness report for comtrade data}
\usage{
createcompletenessreport(rawdata, products = "all",
  inputpath = system.file("templates", package = "tradeflows"),
  template = "completeness.Rmd", outputdir = "reports")
}
\arguments{
\item{rawdata}{a dataframe of tradeflows data}

\item{products}{a vector of product codes}

\item{inputpath}{path of the template, defaults to package internal path}

\item{template}{name of the template file}

\item{outputdir}{name of the output directory relative to getwd()}
}
\description{
You may want to generalise this function, for all template that are
product specific.
}

