% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/report.R
\name{createcountryreport}
\alias{createcountryreport}
\title{Create reports for all products in the given country}
\usage{
createcountryreport(country, template = "allproducts.Rmd",
  outputdir = "reports/countries", ...)
}
\arguments{
\item{country}{}

\item{template}{name of the template file.}

\item{outputdir}{path where report will be saved, relative to the working directory
or an absolute path.}

\item{...}{further arguments passed to \code{\link{createreport}()}
\dontrun{
createcountryreport("China")
}}
}
\description{
The location of the report can be changed by changing the outputdir parameter.
See arguments of \code{\link{createreport}()}
to determine in which folder the template is located.
}

