% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/report.R
\name{creatediscrepancyreport}
\alias{creatediscrepancyreport}
\title{Create a discrepancy report}
\usage{
creatediscrepancyreport(productcode_, reporter_, template = "discrepancy.Rmd",
  outputdir = "reports/discrepancies", toc = FALSE, ...)
}
\arguments{
\item{...}{arguments passed to \code{\link{createreport}()}}

\item{productcode}{vector of product codes}

\item{reporter}{one single country name}
}
\description{
Create a discrepancy report
}
\examples{
\dontrun{
creatediscrepancyreport(440799, "Cameroon", outputdir = "reports/discrepancies")
}
}

