% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/report.R
\name{creatediscrepancyreport}
\alias{creatediscrepancyreport}
\title{Create a discrepancy report}
\usage{
creatediscrepancyreport(productcode, reporter, ...)
}
\arguments{
\item{productcode}{vector of product codes}

\item{reporter}{one single country name}

\item{...}{arguments passed to \code{\link{createproductreport}()}}
}
\description{
Create a discrepancy report
}
\examples{
\dontrun{
creatediscrepancyreport(440799, "Cameroon")
}
}

