% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/report.R
\name{createoverviewreport}
\alias{createoverviewreport}
\title{Create reports for all products in the given reporter country}
\usage{
createoverviewreport(reporter_, productcode_ = NULL, beginyear = 0,
  endyear = 9999, template = "overview_tradevalue.Rmd",
  outputdir = "reports/overview", tableread = "validated_flow_yearly",
  fileprefix = gsub("\\\\..*", "", template), dataonly = FALSE, ...)
}
\arguments{
\item{reporter_}{}

\item{productcode_}{vector of product code, NULL by default to generate the report for all products}

\item{template}{name of the template file.}

\item{outputdir}{path where report will be saved, relative to the working directory
or an absolute path.}

\item{...}{further arguments passed to \code{\link{createreport}()}}
}
\description{
The location of the report can be changed by changing the outputdir parameter.
See arguments of \code{\link{createreport}()}
to determine in which folder the template is located.
}
\examples{
\dontrun{
createoverviewreport("China", beginyear = 2010, endyear = 2013, outputdir = "/tmp")
createoverviewreport("China", beginyear = 2010, endyear = 2013, inputpath = "inst/templates")
}
}

