% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{createoverviewreport}
\alias{createoverviewreport}
\title{Create reports for all products in the given reporter country}
\usage{
createoverviewreport(reporter_, productcode_ = NULL, beginyear = 0,
  endyear = 9999, template = "overviewtradevalue.Rmd",
  outputdir = "reports/overview", tableread = "validated_flow_yearly",
  jfsqlevel = 2, fileprefix = gsub("\\\\..*", "", template),
  dataonly = FALSE, ...)
}
\arguments{
\item{reporter_}{character name of a country reporting data}

\item{productcode_}{vector of product code, NULL by default to generate the report for all products}

\item{template}{name of the template file.}

\item{outputdir}{path where report will be saved, relative to the working directory
or an absolute path.}

\item{jfsqlevel}{integer 1 or 2 the level of jfsq product names to be used, see the internal table classificationitto}

\item{...}{further arguments passed to \code{\link{createreport}()}}
}
\description{
The location of the report can be changed by changing the outputdir parameter.
See arguments of \code{\link{createreport}()}
to determine in which folder the template is located.
}
\details{
jfsqlevel determines the product group levels,
there are approximatley a dozen of product groups under jfsq level 1
and sixty product groups under level 2.
Depending on the level choosen, there will be a dozen or sixty titles in the overview report.
}
\examples{
\dontrun{
createoverviewreport("China", beginyear = 2010, endyear = 2013, outputdir = "/tmp")
createoverviewreport("China", beginyear = 2010, endyear = 2013, inputpath = "inst/templates")
}
}

