% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{estimatequantity}
\alias{estimatequantity}
\title{Estimate quantity}
\usage{
estimatequantity(dtf, price, conversionfactor)
}
\arguments{
\item{dtf}{data frame}

\item{price}{a data frame of unit prices which can be merge with dtf}

\item{conversionfactor}{a data frame of conversion factors which can be
merged with dtf}
}
\description{
Estimate missing quantity
For each trade flow in the given data frame,
compute quantity_cf from the weight using a conversion factor and
comput quantity_up from the trade value using a unit price.
These columns will be available in the data frame output of this function,
but they will not be saved in the validated database.
}
\details{
Split the data frame between flows which have a quantity from those which don't
complete missing quantity from weight when available or from the
tradevalue when the weight is not available
Check upper and lower bounds on price against a table of reference unit prices
In general NA values should be avoided for the
 upper and lower bounds on prices
}

