% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/clean.R
\name{extractconversionfactors}
\alias{extractconversionfactors}
\title{Extract median converion factor at a given geographical aggregation level}
\usage{
extractconversionfactors(dtf, geoaggregation = "region",
  includeqwestimates = TRUE)
}
\arguments{
\item{dtf}{a dataframe containing conversion factors}

\item{geoaggregation}{a character string specifying the regional aggregation
level, "world" to extract world conversion factors,
 "region" to extract regional conversion factors.}

\item{includeqwestimates}{logical TRUE when comtrade quantity and weight
estimates can be included}
}
\description{
Extract median converion factors for the whole world (default)
}

