% Generated by roxygen2 (4.0.2): do not edit by hand
\name{extractconversionfactors}
\alias{extractconversionfactors}
\title{Extract median converion factor at a given geographical aggregation level}
\usage{
extractconversionfactors(dtf, geoaggregation = "World",
  includeqwestimates = TRUE)
}
\arguments{
\item{dtf}{a dataframe containing conversion factors}

\item{geoaggregation}{a character string specifying the regional aggregation
level, a column name in the reportercomtrade table}

\item{includeqwestimates}{logical TRUE when comtrade quantity and weight
estimates can be included}
}
\description{
Extract median converion factors for the whole world (default)
}

