% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{extractconversionfactors}
\alias{extractconversionfactors}
\title{Extract median conversion factor at a given geographical aggregation level}
\usage{
extractconversionfactors(dtf, grouping = c("flow", "regionreporter", "year",
  "unit"), includeqwestimates = TRUE)
}
\arguments{
\item{dtf}{a dataframe containing all trade flows for one product
and their individual conversion factors}

\item{includeqwestimates}{logical TRUE when comtrade quantity and weight
estimates can be included}
}
\description{
Group trade flows by the given geographical aggregation level and
extract the median conversion factors.
}

