% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clean.R
\name{extractconversionfactors}
\alias{extractconversionfactors}
\title{Extract median conversion factor at a given geographical aggregation level}
\usage{
extractconversionfactors(dtf, geoaggregation = "region",
  includeqwestimates = TRUE)
}
\arguments{
\item{dtf}{a dataframe containing all trade flows for one product
and their individual conversion factors}

\item{geoaggregation}{a character string specifying the regional aggregation
level, "world" to extract world conversion factors,
 "region" to extract regional conversion factors.}

\item{includeqwestimates}{logical TRUE when comtrade quantity and weight
estimates can be included}
}
\description{
Group trade flows by the given geo aggregation level and
extract the median conversion factors.
}

