% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/report.R
\name{extractmetadata}
\alias{extractmetadata}
\title{Extract report metadata from trade flow table
This can be used to check that the given trade flow data has only
one reporter and product or the empty string as expected by some reports.}
\usage{
extractmetadata(tfdata)
}
\arguments{
\item{tfdata}{a table of trade flows data}
}
\value{
a list of metadata
}
\description{
Extract report metadata from trade flow table
This can be used to check that the given trade flow data has only
one reporter and product or the empty string as expected by some reports.
}

