% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{extractprices}
\alias{extractprices}
\title{Extract median prices at a given geographical aggregation level}
\usage{
extractprices(dtf, lowercoef = 0.5, uppercoef = 2, grouping = c("flow",
  "regionreporter", "year", "unit"), includeqestimates = TRUE)
}
\arguments{
\item{dtf}{a dataframe containing all trade flows for one product
and their individual prices}

\item{lowercoef}{numeric multiplier of the first quartile to
obtain a lower bound on prices}

\item{uppercoef}{numeric multiplier of the third quartile to
obtain an upper bound on prices}

\item{grouping}{a character vector specifying the grouping variables}

\item{includeqestimates}{logical TRUE when comtrade quantity estimates
can be included}
}
\description{
Group trade flows by the given geographical aggregation level
(region or subregion) and extract the median prices.
Prices depend on the quantity unit. For some products quantity unit changes
from litre to m3 along the years.
}

