% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{findduplicatedflows}
\alias{findduplicatedflows}
\title{Return a dataframe of duplicated flows}
\usage{
findduplicatedflows(dtf)
}
\arguments{
\item{dtf}{a dataframe contiaining trade flows, the funciton
tests for duplicated reportercode, partnercode, productcode, flow, period}
}
\description{
Return a dataframe of duplicated flows
}

