% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{loadcomtradebycode}
\alias{loadcomtradebycode}
\title{load raw data from the COMTRADE API}
\usage{
loadcomtradebycode(productcode, reportercode, year, freq = "A", px = "HS",
  max = 50000, logfile = FALSE)
}
\arguments{
\item{productcode}{code of the product}

\item{reportercode}{geographical area or country code}

\item{year}{a vector of years (maximum 5), or the chain of character "recent"}

\item{freq}{the frequency: "A" Annual, "M" Monthly}

\item{px}{Trade data classification scheme}

\item{max}{maximum records returned}
}
\value{
a dataframe
}
\description{
load a JSON file from the comtrade API in the temp directory
The API is documented at http://comtrade.un.org/data/doc/api/
Converts the file to a dataframe
}
\examples{
\dontrun{
library(tradeflows)
# Load "other sawnwood" in France for the last 5 years available in comtrade
loadcomtradebycode(440799, 251, "recent")
}
}

