% Generated by roxygen2 (4.0.2): do not edit by hand
\name{loadcomtradewithpause}
\alias{loadcomtradewithpause}
\title{Load flows for all countries, with a pause of 1 hour between products}
\usage{
loadcomtradewithpause(products, year = "recent", pause = 3601)
}
\arguments{
\item{year}{a vector of years (maximum 5), or the chain of character "recent"}

\item{productcode}{a vector of product codes}
}
\description{
Load flows in all directions available in comtrade
for a given vector of product codes.
}

