% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{readdbtbl}
\alias{readdbtbl}
\title{[Deprecated] Create a dplyr connector to MySQL}
\usage{
readdbtbl(tableread)
}
\arguments{
\item{tableread}{name of the database table to read}
}
\value{
a dplyr tbl object
}
\description{
\code{readdbtbl} is deprecated, use \code{dplyr::\link{tbl}} instead.
See example below.
}
\details{
[Deprecated] Old description

Return a dplyr tbl object for the given MySQL database table
tbl objects allow lazy operations. See src_mysql.
The issue is that 2 table objects created with this function will
not share the same source.
Therefore they will not be available for a merge within the database.
}
\examples{
\dontrun{
# From 2017 onwards, use
con <- RMariaDB::dbConnect(RMariaDB::MariaDB(), dbname = "tradeflows")
prod <- tbl(con, "raw_comext_product")
RMariaDB::dbDisconnect(con)
# Deprecated, before 2017
readdbtbl("raw_flow_yearly") \%>\%
    group_by(productcode) \%>\%
    summarise(nrow = n()) \%>\%
    arrange(nrow) \%>\%
    kable
}
}
