% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{readdbtbl}
\alias{readdbtbl}
\title{Create a dplyr connector to MySQL}
\usage{
readdbtbl(tableread)
}
\arguments{
\item{tableread}{name of the database table to read}
}
\value{
a dplyr tbl object
}
\description{
Return a dplyr tbl object for the given MySQL database table
tbl objects allow lazy operations. See src_mysql.
The issue is that 2 table objects created with this function will
not share the same source.
Therefore they will not be available for a merge within the database.
}
\examples{
\dontrun{
readdbtbl("raw_flow_yearly") \%>\%
    group_by(productcode) \%>\%
    summarise(nrow = n()) \%>\%
    arrange(nrow) \%>\%
    kable
}
}

