% Generated by roxygen2 (4.0.2): do not edit by hand
\name{renamecolumns}
\alias{renamecolumns}
\title{Rename and select columns}
\usage{
renamecolumns(dtf, sourcedb = "comtrade", destdb = "efi")
}
\arguments{
\item{dtf}{data frame}

\item{sourcedb}{source database, where the data was loaded}

\item{destdb}{destination database, where the data is going to}
}
\description{
Rename and select usefull columns for the
forest products trade flow database.
Renaming is based on a data frame called column_names
which is a matching table of column names for the different
databases.
Rename only columns names that are existing in both
sourcedb and destdb.
}
\examples{
column_names
}

