% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clean.R
\name{sanitycheck}
\alias{sanitycheck}
\title{Check various features of the trade flows data frame}
\usage{
sanitycheck(dtf)
}
\arguments{
\item{dtf}{data frame of trade flows data}
}
\description{
"Import" and "Export" and used to add partner flow information.
Should these characters be different then the user should be warned.
}

