% Generated by roxygen2 (4.0.2): do not edit by hand
\name{setdatabaseconfig}
\alias{setdatabaseconfig}
\title{Sets database connection parameters as global R option tradeflowsDB}
\usage{
setdatabaseconfig()
}
\description{
Database connection parameters
Are storred in the installed version of the package
Run this command to find the configuration file:
system.file("config/databaseconf.R", package="tradeflows")
}

